clk: spacemit: define struct k1_ccu_data
Define a new structure type to be used for describing the OF match data.
Rather than using the array of spacemit_ccu_clk structures for match
data, we use this structure instead.

Move the definition of the spacemit_ccu_clk structure closer to the top
of the file, and add the new structure definition below it.

Shorten the name of spacemit_ccu_register() to be k1_clk_register().

Signed-off-by: Alex Elder <elder@riscstar.com>