clk: spacemit: add reset controller support
Define ccu_reset_data as a structure that contains the constant
register offset and bitmasks used to assert and deassert a reset
control on a SpacemiT CCU.  Add the address of an array of these
structures of that type and a count of the number of elements
in that array to the spacemit_ccu_data structure.  Reset support
is optional for SpacemiT CCUs; that pointer field will be null
for CCUs without any resets.

Define a K1 CCU reset controller type, which (for a CCU with resets)
contains the regmap to be used for the controller, a pointer to the
constant reset data, and an embedded a reset controller structure.

Each reset control is asserted or deasserted by updating bits in
a register.  The bits used are defined by an assert mask and a
deassert mask.  In some cases, one (non-zero) mask asserts reset
and a different (non-zero) mask deasserts it.  Otherwise one mask
is nonzero, and the other is zero.  Either way, the bits in
both masks are cleared, then either the assert mask or the deassert
mask is set in a register to affect the state of a reset control.

Signed-off-by: Alex Elder <elder@riscstar.com>